+++ app.py
@@ -7,6 +7,7 @@ from typing import Optional
 from datetime import datetime, timedelta, timezone
 from fastapi.security import OAuth2PasswordBearer
 from jose import jwt, JWTError, ExpiredSignatureError
+from backend.model import signup_check,signin_check,token_check
 
 app = FastAPI()
 
@@ -71,79 +72,79 @@ pwd_context = CryptContext(schemes = ["bcrypt"], deprecated = "auto")
 
 
 
-# model_確認該email是否已經存在，不存在就寫入資料
-def signup_check(name,email,password):
-	connection = pool2.get_connection()
-	mycursor = connection.cursor(dictionary = True)
+# # model_確認該email是否已經存在，不存在就寫入資料
+# def signup_check(name,email,password):
+# 	connection = pool2.get_connection()
+# 	mycursor = connection.cursor(dictionary = True)
 
-	sql = "SELECT email FROM member WHERE email = %s"
-	val = (email,)
-	mycursor.execute(sql,val)
-	signup_result = mycursor.fetchall()
+# 	sql = "SELECT email FROM member WHERE email = %s"
+# 	val = (email,)
+# 	mycursor.execute(sql,val)
+# 	signup_result = mycursor.fetchall()
 	
-	if len(signup_result) != 0:
-		# 關閉資料庫連線
-		connection.close()
-		return {"ok" : False}  # 資料庫已有資料，無法註冊
-	else:
-		sql_insert = "INSERT INTO member (name, email, password) VALUES (%s, %s, %s)"
-		val_insert = (name,email,password)
-		mycursor.execute(sql_insert,val_insert)
-		connection.commit()
+# 	if len(signup_result) != 0:
+# 		# 關閉資料庫連線
+# 		connection.close()
+# 		return {"ok" : False}  # 資料庫已有資料，無法註冊
+# 	else:
+# 		sql_insert = "INSERT INTO member (name, email, password) VALUES (%s, %s, %s)"
+# 		val_insert = (name,email,password)
+# 		mycursor.execute(sql_insert,val_insert)
+# 		connection.commit()
 
-		# 關閉資料庫連線
-		connection.close()
-		return {"ok" : True}  # 新增資料成功
+# 		# 關閉資料庫連線
+# 		connection.close()
+# 		return {"ok" : True}  # 新增資料成功
 
 
-# model_驗證使用者帳密是否正確
-def signin_check(email,password):
-	connection = pool2.get_connection()
-	mycursor = connection.cursor(dictionary = True)
+# # model_驗證使用者帳密是否正確
+# def signin_check(email,password):
+# 	connection = pool2.get_connection()
+# 	mycursor = connection.cursor(dictionary = True)
 
-	sql = "SELECT email, password, name FROM member WHERE email = %s"
-	val = (email,)
-	mycursor.execute(sql,val)
-	signin_result = mycursor.fetchall()
+# 	sql = "SELECT email, password, name FROM member WHERE email = %s"
+# 	val = (email,)
+# 	mycursor.execute(sql,val)
+# 	signin_result = mycursor.fetchall()
 
-	if len(signin_result) == 0 :
-		connection.close() # 關閉資料庫連線
-		return {"ok" : False, "masseage" : "帳號不存在"}
+# 	if len(signin_result) == 0 :
+# 		connection.close() # 關閉資料庫連線
+# 		return {"ok" : False, "masseage" : "帳號不存在"}
 
-	#pwd_context.verify 是 passlib 提供的方法，用來檢查純文本密碼是否與哈希密碼匹配。
-	elif not pwd_context.verify(password,signin_result[0]["password"]):
-		connection.close() # 關閉資料庫連線
-		return {"ok" : False, "masseage" : "密碼不正確"}
+# 	#pwd_context.verify 是 passlib 提供的方法，用來檢查純文本密碼是否與哈希密碼匹配。
+# 	elif not pwd_context.verify(password,signin_result[0]["password"]):
+# 		connection.close() # 關閉資料庫連線
+# 		return {"ok" : False, "masseage" : "密碼不正確"}
 
-	connection.close() # 關閉資料庫連線
-	return {
-		"ok" : True,
-		"name" : signin_result[0]["name"]
-	}
+# 	connection.close() # 關閉資料庫連線
+# 	return {
+# 		"ok" : True,
+# 		"name" : signin_result[0]["name"]
+# 	}
 	
 
-# model_檢查使用者信箱是否存在
-def token_check(email):
-	connection = pool2.get_connection()
-	mycursor = connection.cursor(dictionary = True)
+# # model_檢查使用者信箱是否存在
+# def token_check(email):
+# 	connection = pool2.get_connection()
+# 	mycursor = connection.cursor(dictionary = True)
 
-	sql = "SELECT id, name, email FROM member WHERE email = %s"
-	val = (email,)
-	mycursor.execute(sql,val)
-	token_result = mycursor.fetchall()
+# 	sql = "SELECT id, name, email FROM member WHERE email = %s"
+# 	val = (email,)
+# 	mycursor.execute(sql,val)
+# 	token_result = mycursor.fetchall()
 
-	connection.close() # 關閉資料庫連線
-	if len(token_result) == 0 :
-		return {"ok" : False, "masseage" : "帳號不存在"}
-	else:
-		return {
-			"ok" : True,
-			"data" : {
-				"id" : token_result[0]["id"],
-				"email" : token_result[0]["email"],
-				"name" : token_result[0]["name"]
-			}
-		}
+# 	connection.close() # 關閉資料庫連線
+# 	if len(token_result) == 0 :
+# 		return {"ok" : False, "masseage" : "帳號不存在"}
+# 	else:
+# 		return {
+# 			"ok" : True,
+# 			"data" : {
+# 				"id" : token_result[0]["id"],
+# 				"email" : token_result[0]["email"],
+# 				"name" : token_result[0]["name"]
+# 			}
+# 		}
 
 # jwt_產出jwt token
 def generate_token(name,email):	
@@ -186,29 +187,30 @@ def verify_token(token):
 async def sign_up(request:Request):
 	# 檢查使用者是否有登入
 	# 用request.json() 來解析request傳來的json資料
-	data = await request.json()
-	name = data.get("name")
-	email = data.get("email")
-	hashed_password = pwd_context.hash(data.get("password") ) #將密碼用hash加密
+	# data = await request.json()
+	# name = data.get("name")
+	# email = data.get("email")
+	# hashed_password = pwd_context.hash(data.get("password") ) #將密碼用hash加密
+	return (data)
 
 	# 檢查使用者是否有註冊過，沒有的話就幫使用者註冊資料
-	signUpCheck = signup_check(name,email,hashed_password)
-
-	if signUpCheck ["ok"] == True:
-		return JSONResponse(
-			status_code = 200,
-			content = {
-				"ok": True
-			}
-		)
-	else:
-		return JSONResponse(
-	        status_code = 400,
-	        content = {
-	            "error": True,
-	            "message": "帳號已存在",
-	        }
-		)
+	# signUpCheck = signup_check(name,email,hashed_password)
+	
+	# if signUpCheck ["ok"] == True:
+	# 	return JSONResponse(
+	# 		status_code = 200,
+	# 		content = {
+	# 			"ok": True
+	# 		}
+	# 	)
+	# else:
+	# 	return JSONResponse(
+	#         status_code = 400,
+	#         content = {
+	#             "error": True,
+	#             "message": "帳號已存在",
+	#         }
+	# 	)
 
 #/api/auth 登入會員帳戶，成功的話取得token
 @app.put("/api/auth")
@@ -218,7 +220,7 @@ async def sign_in(request:Request):
 	data = await request.json()
 	email = data.get("email")
 	password = data.get("password")
-	
+
 	#檢查帳密是否正確
 	signin_result = signin_check(email,password)
